requirements.txt
# Gemini API
google-generativeai==0.8.3
langchain==0.3.14
langchain-google-genai==2.0.7

# Backend
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4

# Data & ML
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
torch==2.5.1

# Async
aiohttp==3.11.11

# Utilities
python-dotenv==1.0.1
httpx==0.28.1